Move stuff, add secondary button
